+++ index.html
@@ -12,5 +12,24 @@
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
+    <script>
+      const matcher = window.matchMedia("(prefers-color-scheme: light)");
+
+      const lightSchemeIcon = document.querySelector("link#light-scheme-icon");
+      const darkSchemeIcon = document.querySelector("link#dark-scheme-icon");
+
+      function onUpdate() {
+        if (matcher.matches) {
+          lightSchemeIcon.remove();
+          document.head.append(darkSchemeIcon);
+        } else {
+          darkSchemeIcon.remove();
+          document.head.append(lightSchemeIcon);
+        }
+      }
+
+      matcher.addListener(onUpdate);
+      onUpdate();
+    </script>
   </body>
 </html>
