+++ src/components/Portfolio/MainPanelSkeleton.tsx
