+++ src/pages/Portfolio/PortfolioPage.tsx
@@ -1,11 +1,9 @@
 import { OverviewErrorFallback } from "@components/Dashboard/errorFallback/OverviewErrorFallback";
 import ChartsPanel from "@components/Portfolio/Charts/ChartsPanel";
 import ChartsPanelSkeleton from "@components/Portfolio/Charts/skeletons/ChartsPanelSkeleton";
-import MainPanelSkeleton from "@components/Portfolio/Charts/skeletons/MainPanelSkeleton";
+import MainPanelSkeleton from "@components/Portfolio/MainPanelSkeleton";
 import MainPanel from "@components/Portfolio/MainPanel";
-import PortfolioHoldingAddDialog from "@components/Portfolio/PortfolioHoldings/PortfolioHoldingAddDialog";
 import { AsyncBoundary } from "@components/common/AsyncBoundary";
-import { useState } from "react";
 import styled from "styled-components";
 import BasePage from "../BasePage";
 
@@ -19,19 +17,13 @@ export default function PortfolioPage() {
   //   eventTypeName: "portfolioDetails",
   // });
 
-  const [isAddHoldingDialogOpen, setIsAddHoldingDialogOpen] = useState(false);
-
-  const onAddHoldingButtonClick = () => {
-    setIsAddHoldingDialogOpen(true);
-  };
-
   return (
     <BasePage>
       <Container>
         <AsyncBoundary
           ErrorFallback={OverviewErrorFallback}
           SuspenseFallback={<MainPanelSkeleton />}>
-          <MainPanel onAddHoldingButtonClick={onAddHoldingButtonClick} />
+          <MainPanel />
         </AsyncBoundary>
 
         <AsyncBoundary
@@ -40,11 +32,6 @@ export default function PortfolioPage() {
           <ChartsPanel />
         </AsyncBoundary>
       </Container>
-
-      <PortfolioHoldingAddDialog
-        isOpen={isAddHoldingDialogOpen}
-        onClose={() => setIsAddHoldingDialogOpen(false)}
-      />
     </BasePage>
   );
 }
